Providing more detailed error when the put API fails (#257)
* s3 0.29.0 (#253)

* cargo: Move `tokio-stream` create behind the `with-tokio` feature (#248)

`tokio-stream` has a dependency on `tokio` and is only used within the
`request` module, which depends on the `with-tokio` feature.  As such,
in order to completely prevent `tokio` from appearing in the crate
graph, this crate must be marked optional and enabled when `with-tokio`
is requested, omitted otherwise.

Also includes two drive-by typo-fixes in the features section of the
readme.

* Fixup #248

* Migrate from `chrono` to `time` to solve RUSTSEC-2020-0159 (#250)

`chrono` still hasn't found a solution to [RUSTSEC-2020-0159] whereas
`time` already solved its vulnerability to [RUSTSEC-2020-0071] by hiding
the affected functionality behind a cfg flag (not to be confused with a
`feature`, such `cfg`s can only be enabled through `RUSTFLAGS`).  At the
same time `chrono` is a superset of `time` even though this crate hardly
uses any functionality of it: only UTC time is needed which does not
suffer from aforementioned local time vulnerabilities.

[RUSTSEC-2020-0071]: https://rustsec.org/advisories/RUSTSEC-2020-0071
[RUSTSEC-2020-0159]: https://rustsec.org/advisories/RUSTSEC-2020-0159

Co-authored-by: Drazen Urch <drazen@urch.eu>

* Update Makefile

* Closes #245

* Bump aws-region

* s3 0.29.0

Co-authored-by: Marijn Suijten <marijns95@gmail.com>

* Provide better error information

Co-authored-by: Drazen Urch <drazen@urch.eu>
Co-authored-by: Marijn Suijten <marijns95@gmail.com>